re-organize utilities and more cleanup
