+++ src/__tests__/testUtils.ts
@@ -1,7 +1,8 @@
+import * as path from "path";
 import { assert } from "vitest";
 import { isEqual } from "lodash";
-import * as path from "path";
-import { ComponentDoc, parse, ParserOptions } from "../parser";
+import { parse } from "../parser";
+import type { ComponentDoc, ParserOptions } from "../parser/types";
 
 export interface ExpectedComponents {
   [key: string]: ExpectedComponent;
@@ -60,7 +61,6 @@ export function checkComponent(
   assert.equal(
     actual.length,
     expectedComponentNames.length,
-    // tslint:disable-next-line:max-line-length
     `The number of expected components is different - \r\nexpected: ${expectedComponentNames}, \r\nactual: ${actual.map(
       (i) => i.displayName
     )}`
@@ -87,17 +87,12 @@ export function checkComponent(
     }
 
     if (componentDoc.description !== expectedComponentDescription) {
-      // tslint:disable-next-line:max-line-length
       errors.push(
         `${compName} description is different - expected: '${expectedComponentDescription}', actual: '${componentDoc.description}'`
       );
     }
 
-    if (
-      propNames.length !== expectedPropNames.length &&
-      exactProperties === true
-    ) {
-      // tslint:disable-next-line:max-line-length
+    if (propNames.length !== expectedPropNames.length && exactProperties) {
       errors.push(
         `Properties for ${compName} are different - expected: ${
           expectedPropNames.length
@@ -114,7 +109,6 @@ export function checkComponent(
         errors.push(`Property '${compName}.${expectedPropName}' is missing`);
       } else {
         if (expectedProp.type !== prop.type.name) {
-          // tslint:disable-next-line:max-line-length
           errors.push(
             `Property '${compName}.${expectedPropName}' type is different - expected: ${expectedProp.type}, actual: ${prop.type.name}`
           );
@@ -125,7 +119,6 @@ export function checkComponent(
             : expectedProp.description;
         if (expectedDescription !== prop.description) {
           errors.push(
-            // tslint:disable-next-line:max-line-length
             `Property '${compName}.${expectedPropName}' description is different - expected: ${expectedDescription}, actual: ${prop.description}`
           );
         }
@@ -138,7 +131,6 @@ export function checkComponent(
           expectedParentFileName !== prop.parent.fileName
         ) {
           errors.push(
-            // tslint:disable-next-line:max-line-length
             `Property '${compName}.${expectedPropName}' parent fileName is different - expected: ${expectedParentFileName}, actual: ${prop.parent.fileName}`
           );
         }
@@ -146,7 +138,6 @@ export function checkComponent(
           expectedProp.required === undefined ? true : expectedProp.required;
         if (expectedRequired !== prop.required) {
           errors.push(
-            // tslint:disable-next-line:max-line-length
             `Property '${compName}.${expectedPropName}' required is different - expected: ${expectedRequired}, actual: ${prop.required}`
           );
         }
@@ -159,20 +150,17 @@ export function checkComponent(
           expectedDefaultValue !== actualDefaultValue
         ) {
           errors.push(
-            // tslint:disable-next-line:max-line-length
             `Property '${compName}.${expectedPropName}' defaultValue is different - expected: ${expectedDefaultValue}, actual: ${actualDefaultValue}`
           );
         }
         const exptectedRaw = expectedProp.raw;
         if (exptectedRaw && exptectedRaw !== prop.type.raw) {
-          // tslint:disable-next-line:max-line-length
           errors.push(
             `Property '${compName}.${expectedPropName}' raw value is different - expected: ${exptectedRaw}, actual: ${prop.type.raw}`
           );
         }
         const expectedValue = expectedProp.value;
         if (expectedValue && !isEqual(expectedValue, prop.type.value)) {
-          // tslint:disable-next-line:max-line-length
           errors.push(
             `Property '${compName}.${expectedPropName}' value is different - expected: ${JSON.stringify(
               expectedValue
@@ -193,7 +181,6 @@ export function checkComponent(
   }
   const ok = errors.length === 0;
   if (!ok) {
-    // tslint:disable-next-line:no-console
     console.log(JSON.stringify(actual, null, 4));
   }
