+++ src/parser/utilities/formatTag.ts
@@ -1,0 +1,9 @@
+import * as ts from "typescript";
+
+export function formatTag(tag: ts.JSDocTagInfo) {
+  let result = "@" + tag.name;
+  if (tag.text) {
+    result += " " + ts.displayPartsToString(tag.text);
+  }
+  return result;
+}
