+++ src/parser/utilities/getPropertyName.ts
@@ -1,0 +1,16 @@
+import * as ts from "typescript";
+
+export function getPropertyName(
+  name: ts.PropertyName | ts.BindingPattern
+): string | null {
+  switch (name.kind) {
+    case ts.SyntaxKind.NumericLiteral:
+    case ts.SyntaxKind.StringLiteral:
+    case ts.SyntaxKind.Identifier:
+      return name.text;
+    case ts.SyntaxKind.ComputedPropertyName:
+      return name.getText();
+    default:
+      return null;
+  }
+}
